# solver: return prev solving result for recursive cases