We achieve this by using map[string]map[string]struct{} instead of
map[string]struct{}. The nested maps are the resolveType results for
a given type.

We start by assigning a nil map so we don't fall into the infinite
recursion. When resolveTypeNoLateStaticBinding returns we record
resolving results into the same map key. If we hit the resolving loop,
we return previously computed value (if any).

This does affect the performance by a ~5%:

	name                     old time/op    new time/op    delta
	ExprType/simplevar-8       2.15µs ± 2%    2.27µs ± 1%  +5.43%  (p=0.008 n=5+5)
	ExprType/f1call-8          5.06µs ± 1%    5.22µs ± 1%  +3.23%  (p=0.008 n=5+5)
	ExprType/f3call-8          3.78µs ± 1%    3.92µs ± 2%  +3.49%  (p=0.008 n=5+5)
	ExprType/m4call-8          7.22µs ± 2%    7.52µs ± 1%  +4.19%  (p=0.008 n=5+5)
	ExprType/newpropfetch-8    5.88µs ± 1%    6.04µs ± 0%  +2.66%  (p=0.016 n=5+4)

(Benchmarks are from the linttest bench_test.go suite.)

Fixes #553

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>